feat: rename submodule vector to qnewton
