+++ src/polykin/math/roots/qnewton.py
